Move ui folder, remove raxi:eventemitter, remove streams reducer.
