+++ .meteor/packages
@@ -22,4 +22,3 @@ static-html
 accounts-base
 session
 accounts-password
-raix:eventemitter
