+++ .meteor/versions
@@ -55,7 +55,6 @@ omega:meteor-desktop-localstorage@0.0.11
 omega:meteor-desktop-watcher@0.3.0
 ordered-dict@1.0.9
 promise@0.8.8
-raix:eventemitter@0.1.3
 random@1.0.10
 rate-limit@1.0.6
 reactive-dict@1.1.8
