--- imports/redux/actions.js
@@ -1,1 +1,0 @@
-// TODO: combine actions
