--- imports/redux/reducers.js
@@ -1,1 +1,0 @@
-// TODO: combine reducers
