+++ imports/ui/components/editor/components/localEditorComponent.js
