+++ imports/ui/routes/register/index.js
